pyproject.toml
[tool.poetry]
name = "schemathesis"
version = "1.7.0"
description = "Hypothesis strategies for Open API / Swagger schemas"
keywords = ["pytest", "hypothesis", "openapi", "swagger", "testing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Framework :: Pytest",
    "Framework :: Hypothesis",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development :: Testing",
]
authors = ["Dmitry Dygalo <dmitry.dygalo@kiwi.com>"]
readme = "README.rst"
repository = "https://github.com/kiwicom/schemathesis"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
attrs = "^19.1"
hypothesis = ">=5.15,<6.0"
hypothesis_jsonschema = ">=0.11.1,<1.0"
hypothesis_graphql = ">=0.3"
jsonschema = "^3.0.0"
pytest = ">4.6.4"
pyyaml = "^5.1"
pytest-subtests = ">=0.2.1,<1.0"
requests = "^2.22"
click = "^7.0"
importlib_metadata = { version = "^1.1", python = "<3.8" }
werkzeug = ">0.16.0"
junit-xml = "^1.9"

[tool.poetry.dev-dependencies]
coverage = "^4.5"
pytest = ">4.6.4"
aiohttp = "^3.6"
pytest-mock = "^1.11.0"
pytest-asyncio = "^0.11.0"
pytest-xdist = "^1.30"
typing_extensions = "^3.7"
flask = "^1.1"
sphinx = "^3.0.3"

[tool.poetry.plugins]
pytest11 = {schemathesis = "schemathesis.extra.pytest_plugin"}

[tool.poetry.scripts]
schemathesis = "schemathesis.cli:schemathesis"

[tool.black]
line-length = 120
target_version = ["py37"]

[tool.isort]
# config compatible with Black
line_length = 120
multi_line_output = 3
default_section = "THIRDPARTY"
include_trailing_comma = true
known_first_party = "schemathesis"
known_third_party = ["_pytest", "aiohttp", "attr", "click", "flask", "graphene", "graphql", "graphql_server", "hypothesis", "hypothesis_graphql", "hypothesis_jsonschema", "jsonschema", "junit_xml", "packaging", "pytest", "pytest_subtests", "requests", "schemathesis", "urllib3", "werkzeug", "yaml"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
